Move logic code to
